+++ CHANGELOG.md
@@ -8,11 +8,11 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 ### Added
 - Max-level provenance and Software Bill Of Materials (SBOM) to the Docker images pushed to Docker Hub
 - ACMG VUS Bayesian score / temperature on case reports
+- Button to filter and download case individuals/samples from institute's caseS page
 ### Changed
 - On variant page, RefSeq transcripts panel, truncate very long protein change descriptions
 ### Fixed
 - UCSC hg38 links are updated
-
 
 ## [4.93.1]
 ### Fixed